feat(nest): move to postgreSQL
- Move from rethinkDB to postgres
- Remove database-specific code from client
- Move prefix and locale queries into callables

BREAKING CHANGES:
Signature (and functionality) of NestClient changed.
Prefixes and locales need to be supplied as a constant or callable.

Database shifted from RethinkDB to PostgreSQL.
No migration assistance, as the bot is still early in development.

Any cogs using the old DB wrapper have to be changed to use asyncpg.